requirements.txt
Django==1.5.1
Fabric==1.6.1
Markdown==2.3.1
Pygments==1.6
South==0.8.1
argparse==1.2.1
dj-database-url==0.2.1
django-braces==1.0.0
django-crispy-forms==1.3.2
django-debug-toolbar==0.9.4
django-easy-maps==0.8.1
django-extensions==1.1.1
django-filter==0.6
django-grappelli==2.4.5
django-guardian==1.1.1
django-model-utils==1.4.0
django-notification==1.0
django-oauth2-provider==0.2.6
django-social-auth==0.7.23
django-tastypie==0.9.15
djangorestframework==2.3.6
easy-thumbnails==1.3
geopy==0.95.1
httplib2==0.8
ipython==0.13.2
mimeparse==0.1.3
oauth2==1.5.211
paramiko==1.10.1
path.py==4.3
psycopg2==2.5.1
pycrypto==2.6
python-dateutil==2.1
python-openid==2.2.5
shortuuid==0.3
six==1.3.0
wsgiref==0.1.2
